tests/wpt/web-platform-tests/html/browsers/offline/application-cache-api/api_update_error.html

<!DOCTYPE HTML>
<html manifest="../resources/manifest/clock.manifest">
  <head>
    <title>Offline Application Cache - API_update_error</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <ol>
      <li>Remove the manifest file (manifest/clock.manifest) from the server.</li>
      <li>Refresh the page, then calling update() will throw InvalidStateError exception.</li>
    </ol>

    <div id="log"></div>

    <script>
      var cache = window.applicationCache;
      test(function() {
        assert_throws("INVALID_STATE_ERR", function(){cache.update();});
      }, "INVALID_STATE_ERR error test")
    </script>

  </body>
</html>
